[wlm] Moves LmiUtils.inferLmiEngine() into separate class
